Fix index-as-message bug in `assertArray`
As `assertArray` previously called the assertion
function in a `forEach(assertion)`, the index was
passed as the custom message. This commit fixes
this bug by passing down the `message` argument
to the assertion function.

A `@ts-expect-error` comment was added to due to
"Assertions require every name in the call target
to be declared with an explicit type
annotation.".